model_service/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
